requirements.txt
attrs==22.2.0
anytree==1.16.0
GitPython==3.1.31
requests==2.28.0
